Harmonize states between artist/library view
